dir & first reset

+++ reset.py
@@ -1,11 +1,17 @@
 from general import *
 
-Paper , num_paper = Connect_To_Paper()
-
-for x in range(num_paper):
-    Paper.seek((1+x)*size_paper)                               #拿刪除名
-    delete_paper_name = Paper.read((int)(size_paper/2))
-    Paper.seek((1+x)*size_paper)
-    delete_paper_name = Paper.read(strlen(delete_paper_name,size_paper))  #拿刪除名
-    os.remove(delete_paper_name)
+if not os.path.isdir("paper"):
+    os.mkdir("paper")
+    print("\033[1;33;40m已新增paper資料夾\033[0m")
+    Paper = open("./paper/Paper" , "w+" )
+    print("\033[1;33;40m已新增Paper檔案\033[0m")
+else:
+    Paper , num_paper = Connect_To_Paper()
+    for x in range(num_paper):
+        Paper.seek((1+x)*size_paper)                                    #拿刪除名
+        delete_paper_name = Paper.read((int)(size_paper/2))
+        Paper.seek((1+x)*size_paper)
+        delete_paper_name = Paper.read(strlen(delete_paper_name,size_paper))
+        delete_paper_name = "./paper/" + delete_paper_name              #拿刪除名
+        os.remove(delete_paper_name)
 update_num( Paper , 0 )
